enable debug implicit util to show substrait content when use spark-shell.
